Move comparison functions to own file
- Move the frame comparison functions in ./index.js to own file. Update
other files as needed.
